adding regexp to sql when searching blocks
This is to evade the problem in which searching for Diamond Ore brought up Quartz Stairs.
In SQL syntax, %SEARCH_STRING%, % matches everything, even digits. Diamond ore is 56, Quartz Stairs 156
The provided example negates digits and some special symbols for quantity and after-id values to trim the results to the proper ones.
Note: This so far reduces a lot results when searching for both a keyword and a block, rarely used combination I hope.